frequency<- {zoo}	R Documentation

Replacing the Index of Objects

Description

Generic function for replacing the frequency of an object.

Usage

frequency(x) <- value

Arguments

x
an object.

value
a frequency.

Details

frequency<- is a generic function for replacing (or assigning) the frequency of an object. Currently, there is a "zooreg" and a "zoo" method. In both cases, the value is assigned to the "frequency" of the object if it complies with the index(x).

See Also

zooreg, index

Examples

z <- zooreg(1:5)
z
as.ts(z)
frequency(z) <- 3
z
as.ts(z)

[Package zoo version 1.8-10 Index]